prod/requirements.txt
torch
torchvision

streamlit>=1.35
numpy
pandas
scikit-learn
Pillow
matplotlib
seaborn
requests
